requirements.txt
requests
python-dotenv
gradio
pypdf
PyPDF2
openai
openai-agents